% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctedfit.R
\name{correctedfit}
\alias{correctedfit}
\title{Fit correction for large SEM}
\usage{
correctedfit(model)
}
\arguments{
\item{model}{a fitted lavaan or lsttheory model}
}
\description{
This function calculates a corrected \eqn{\chi^2}-value based on Yuan et al. (2015). 
A corresponding p-value and RMSEA value are computed based on the corrected \eqn{\chi^2} value. 
This correction is intended for SEM with many manifest variables. The \eqn{\chi^2}-value is multiplied 
with an empirical correction factor \eqn{e = (N - (2.381 + 0.361 \cdot p + 0.006 \cdot q) ) / (N - 1)}, 
where p is the total number of manifest variables and q is the number of free parameters.
}
\references{
Yuan, K.-H., Tian, Y., & Yanagihara, H. (2015). Empirical correction to the likelihood ratio statistic for structural equation modeling with many variables. Psychometrika, 80(2), 379–405.
}
\author{
Julia Norget
}
