% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_models_es.R
\docType{data}
\name{d_lst_es}
\alias{d_lst_es}
\title{Dataset d_lst_es.}
\format{A data frame with 1000 rows and 27 variables}
\description{
A simulated dataset to test experience sampling LST models. The variables are:
}
\details{
\itemize{
  \item y11 
  \item ...
  \item ...
  \item y39
}
}
\keyword{datasets}
