% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{d_multitraitmultistate}
\alias{d_multitraitmultistate}
\title{Dataset d_multitraitmultistate.}
\format{
A data frame with 1000 rows and 8 variables
}
\description{
A simulated dataset. The variables are:
}
\details{
\itemize{
  \item y11.
  \item y21.
  \item y12.
  \item y22.
  \item y13.
  \item y23.
  \item y14.
  \item y24.
}
}
\keyword{datasets}
