% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{happy}
\alias{happy}
\title{Dataset happy}
\format{
A data frame with 425 rows and 50 variables
}
\description{
Restructured wide-format data from Weiss et al. (2021), containing the items 
happiness ("How happy do you feel at the moment") and 
life satisfaction ("How satisfied are you with your life at the moment?") on five-point Likert scales.
Data was assessed 5 times a day on 5 days during an experience-sampling study.
The original dataset is available online at: https://osf.io/kwp6n/ 
The variables are:
}
\details{
\itemize{
  \item happy_1.
  \item satisfaction_1.
  \item happy_2.
  \item satisfaction_2.
  \item \cdot
  \item happy_25.
  \item satisfaction_25.
}
}
\keyword{datasets}
