% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_models_es.R
\name{lst_models_es}
\alias{lst_models_es}
\title{Estimate latent state-trait models for experience sampling data}
\usage{
lst_models_es(
  traitmodel,
  ntimepoints,
  data,
  nperiods = 1,
  equiv = "invar",
  ar = TRUE,
  ...
)
}
\arguments{
\item{traitmodel}{character. Can be one of c("singletrait", "day-specific", "indicator-specific", "day-and-indicator-specific")}

\item{ntimepoints}{integer. Number of measurement occasions}

\item{data}{data.frame.}

\item{nperiods}{integer. Number of periods (trait periods, zeta periods, and epsilon periods).}

\item{equiv}{character. Equivalence assumption. Can be one of c("invar", "period.invar", "free")}

\item{ar}{logical. Should autoregressive effects be included}

\item{...}{further arguments passed to lower level functions}
}
\value{
object of class lstmodel
}
\description{
This function can be used to estimate various latent state-trait models (LST models)
for experience sampling data. It is based on the revised version of the LST theory presented in 
Steyer, Mayer, Geiser & Cole (2015) and on Eid et al. (2017). The function is a wrapper for
lst_models_es_common_trait() and lst_models_es_indicator_specific_trait().
}
\examples{
m1 <- lst_models_es(traitmodel="singletrait", ntimepoints=9, 
data=d_lst_es, nperiods=3, ar=FALSE, equiv="invar")
}
\references{

}
