% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsttheory_shiny.R
\name{lstesGUI}
\alias{lstesGUI}
\title{Shiny interface for lst_models_es()}
\usage{
lstesGUI()
}
\description{
This function calls a shiny interface for LST-R models with experience sampling data.
}
\author{
Julia Norget
}
