% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsttheory_es.R
\name{lsttheory_es}
\alias{lsttheory_es}
\title{Estimate latent state-trait models for experience sampling data}
\usage{
lsttheory_es(
  model,
  ntimepoints,
  nperiods,
  data,
  missing = "fiml",
  la_t_equiv = NULL,
  la_o_equiv = NULL,
  la_s_equiv = NULL,
  vzeta_eqiv = NULL,
  veps_equiv = NULL,
  vtheta_equiv = NULL,
  nu_equiv = NULL,
  alpha_equiv = NULL,
  mtheta_equiv = NULL,
  ...
)
}
\arguments{
\item{model}{integer or character.}

\item{ntimepoints}{integer. The total number of measurement occasions on which data was collected.}

\item{nperiods}{integer. The number of periods (e.g. days or weeks) on which data was collected.}

\item{data}{a data.frame. This data frame contains the observed variables, sorted by time t and then 
by indicator i, i.e., Y11, Y21, Y31, ... Y12, Y22, Y32 ... Y15, Y25, Y35 ... etc.}

\item{missing}{lavaan Option. Default is set to "fiml" instead of the standard default option "listwise" with lavaan.
See \code{\link[lavaan]{lavOptions}} for details.}

\item{la_t_equiv}{Character. Factor loadings of the latent trait. 
Can be one of \code{c("one", "period.invar", "free")}.}

\item{la_o_equiv}{Character. Factor loadings of the occasion factor (OCC). 
Can be one of \code{c("one", "time.invar", "period.invar", "free")}.
In models without autoregression this corresponds to the factor loadings 
of the state variables (models with a single trait or period-specific traits) or state residual variables (models with indicator-specific traits).}

\item{la_s_equiv}{Character. Autoregression between occasion factors. Can be one of \code{c("zero", "time.invar", "period.invar", "free")}.}

\item{vzeta_eqiv}{Character. Variances of the state residual (zeta) variables. Can be one of \code{c("time.invar", "period.invar", "free")}.}

\item{veps_equiv}{Character. Variances of the residual (epsilon) variables. 
Can be one of \code{c("invar",  "time.invar", "indicator.invar", "period.invar", "free")}.}

\item{vtheta_equiv}{Character. Variances of the latent trait. Can be one of \code{c("invar","indicator.invar", "free")}.}

\item{nu_equiv}{Character. Intercepts of the indicators. Can be one of \code{c("zero","period.invar", "free")}.}

\item{alpha_equiv}{Character. Intercepts of the latent states. 
Only relevant for models with a single trait or period-specific traits.
Can be one of \code{c("zero","period.invar", "free")}.}

\item{mtheta_equiv}{Means of the latent traits. Character. Can be one of \code{c("invar","indicator.invar", "free")}.}

\item{...}{Further arguments passed to lower-level functions}
}
\description{
Estimate latent state-trait models for experience sampling data
}
\author{
Julia Norget
}
