% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmLSTrf.R
\name{mmLSTrf}
\alias{mmLSTrf}
\title{Multi-Method Latent-State-Trait Random-Fixed Models}
\usage{
mmLSTrf(
  data,
  nSit = 2,
  nTime = 2,
  nMth = 2,
  structural = "TF",
  includeOMF = TRUE,
  lat.cov = list(TFcov = FALSE, OFcov = FALSE, TMFcov = FALSE, OMFcov = FALSE),
  meanstructure = FALSE,
  meas.invar = "time.invar",
  equiv.ass = list(TF = "cong", OF = "cong", OMF = "cong"),
  addsyntax = "",
  ...
)
}
\arguments{
\item{data}{Data frame.\cr 
This data frame contains only the indicator variables which will be used to 
fit the MM-LST-RF model. For more information about the required data 
structure refer to the function details.
The first indicator & the first method will be used as reference indicator & 
reference method across occasions and fixed situations.}

\item{nSit}{Integer. Default = 2.\cr 
Number of fixed situations. Must be >= 2.
Used to calculate the number of Trait Factors (TF).}

\item{nTime}{Integer. Default = 2.\cr 
Number of time-points. Must be >= 2. 
Used to calculate the number of Occasion Factors (OF).}

\item{nMth}{Integer. Default = 2.\cr 
Number of methods. Must be >= 2.
Used to calculate the number of Trait-Method Factors (TMF) & Occasion-Method
Factors (OMF).}

\item{structural}{Character string. Default = \code{"TF"}.\cr
Adds structural part to the model. Can be one of the following:
\itemize{
\item \code{"none"} estimates only the measurement model.
\item \code{"TF"} models Person x Situation interactions. The TF from fixed 
situation s = 1 serves as reference to be contrasted against. \code{TFcov} will be 
set to \code{FALSE}.
\item \code{"TMF"} models Method x Situation interactions. The TMF from fixed 
situation s = 1 serve as references to be contrasted against. Only TMF of the 
same indicator i and method m will be contrasted across situations.\code{TFcov} 
will be set to \code{FALSE}.
\item \code{"both"} models both Person x Situation & Method x Situation interactions
modeled. \code{TFcov} and \code{TMFcov} will be set to \code{FALSE}.
}}

\item{includeOMF}{Logical. Default = \code{TRUE}.\cr 
If \code{TRUE}, OMF will be estimated. 
If \code{FALSE}, no OMF will be estimated. 
Can be used when only small or no occasion-method effects are expected.}

\item{lat.cov}{List. Default = \code{FALSE}.\cr 
Specifies covariances between latent variables.
\itemize{
\item \code{TFcov}. If \code{TRUE}, covariances between all TF will be estimated. 
Will automatically be set to \code{FALSE} if \code{structural = "TF" || "both"}.
If \code{FALSE}, no covariances between TF will be estimated.
\item \code{OFcov}. If \code{TRUE}, covariances of OF with their respective 
counterparts in other fixed situations will be estimated (e.g., Cv(O21,
O_22), Cv(O_21,O_23) and Cv(O_22, O_23)). 
If FALSE, no covariances between OF will be estimated.
\item \code{TMFcov}. If \code{TRUE}, covariances between all TMF will be estimated. 
Will automatically be set to FALSE if \code{structural = "TMF" || "both"}.
If \code{FALSE}, no covariances between TMF will be estimated.
\item \code{OMFcov}. If \code{TRUE}, covariances of OMF with their respective 
counterparts in other fixed situations will be estimated.
If \code{FALSE}, no covariances between OMF will be estimated.
}
It is recommended to estimate covariances sparingly (especially for TMF) as 
this quickly inflates the amount of free model parameters. 
For the estimation of only a select number of covariances, specifications 
can be passed to the function via the \code{"addsyntax"} argument.}

\item{meanstructure}{Logical. Default = \code{FALSE}.\cr 
If \code{TRUE}, model will be estimated with a meanstructure. 
If \code{FALSE}, model will be estimated without a meanstructure.}

\item{meas.invar}{Character string. Default = \code{"time.invar"}.\cr
Measurment invariance assumptions across groups. Can be one of the following:
\itemize{
\item \code{"time.invar"} = assumes time invariance of indicator loadings + intercepts
across occasions.
\item \code{"metric.m"} assumes time invariance & equal loadings across methods. 
\item \code{"metric.s"} assumes time invariance & equal loadings across fixed situations.
\item \code{"metric.b"} assumes time invariance & equal loadings across methods & fixed 
situations. 
\item \code{"scalar.m"} assumes time invariance & equal loading + intercepts across 
methods.  
\item \code{"scalar.s"} assumes time invariance & equal loading + intercepts across 
fixed situations. 
\item \code{"scalar.b"} assumes time invariance & equal loading + intercepts across 
methods & fixed situations.
\item \code{"residual.m"} assumes time invariance & equal loading + intercepts + error 
variances across methods.  
\item \code{"residual.s"} assumes time invariance & equal loading + intercepts + error 
variances across fixed situations.
\item \code{"residual.b"} assumes time invariance & equal loading + intercepts + error 
variances across methods & fixed situations.
}}

\item{equiv.ass}{List. Default = \code{"cong"}.\cr
Equivalence assumptions for latent variables.
For TF, OF & OMF the following equivalence assumptions can be made separately 
for each latent variable:
\itemize{
\item \code{"cong"} for congeneric measures makes no restrictions apart from setting the
loading of the first indicator of the latent variable to 1 and the intercept 
to 0 for identification purposes.
\item \code{"ess.equiv"} for essentially equivalent measures assumes equal loadings of
all indicators of the latent variable.
}
For TF the following additional equivalence assumptions can be made:
\enumerate{
\item \code{"equiv"} for equivalent measures assumes equal loadings + intercepts of all 
indicators of TF.
\item \code{"ess.par"} for essentially parallel measures assumes equal loadings + error 
variances of all indicators of TF.
\item \code{"par"} for parallel measures assumes equal loadings, intercepts + error 
variances of all indicators of TF.
} 
No equivalence assumptions need to be specified for TMF since they are defined 
as residual latent variables and all loadings are automatically fixed to one 
due to the combination of identification requirements and time invariance.}

\item{addsyntax}{Character string.\cr
Optional argument where additional model specifications, such as covariances
between specific TMFs and TFs, can be passed to the function and will 
automatically be added to the generated lavaan syntax. 
For more information about the used variable notation refer to the function 
details.}

\item{...}{Further arguments passed to \code{lavaan::sem()}.}
}
\value{
Object of class mmLSTrf.
}
\description{
A function to estimate Multi-Method Latent-State-Trait Random-Fixed SEM 
models based on the MM-LST-RF model developed by Hintz, Geiser & Shiffman, 
2019.
}
\details{
\strong{Data Structure:}\cr
The order of the indicators Y_imts (i = indicator, m = method, t = time-point
/occasion, s = fixed situation) need to match the following format:
Y_1111, Y_2111, Y_1211, Y_2211, Y_1121, Y_2121, Y_1221, Y_2221, Y_1112, etc.
This format reflects the order of indicator variables in a path diagram where
indicators are first grouped by fixed situations, within those they are then
grouped by occasions and within those they are lastly grouped by methods. 
The resulting nested structure has indicators nested within methods, nested
within occasions, nested within fixed situations.

The number of fixed situations, occasions (time-points), methods and number 
of indicators can take any value >= 2.
However, the model is restricted to an equal number of indicators for each 
method, occasion and fixed situation. For example, if method A has 3 
indicators, method B must also have 3 indicators. Accordingly, if 
there are 2 methods with 3 indicators each, every  occasion will have 6 
indicators (3 from each method).  
In the same vein, the number of methods needs to be equal across occasions 
and the number of occasions needs to be equal across fixed situations. 
Lastly, only one Trait Factor is estimated per fixed situation.

\strong{Syntax notation:}\cr
Example for Y_2132 with i = 2, m = 1, t = 3, s = 2 (covariances with s = 3)
\tabular{lll}{
\strong{Variable / Parameter}           \tab \strong{Notation}  \tab \strong{Example}   \cr
Trait factors (TF)                      \tab Ts                 \tab T2                 \cr
Occasion factors (OF)                   \tab Ots                \tab O32                \cr
Trait-method factors (TMF)              \tab TMims              \tab TM212              \cr
Occasion-method factors (OMF)           \tab OMmts              \tab OM132              \cr
TF mean                                 \tab M_Ts               \tab M_T2               \cr
TF variance                             \tab V_Ts               \tab V_T2               \cr
OF variance                             \tab V_Ots              \tab V_O32              \cr
OMF variance                            \tab V_OMmts            \tab V_OM132            \cr
TMF variance                            \tab V_TMims            \tab V_TM212            \cr
Observed variable variance              \tab V_Y_imts           \tab V_(variable name)  \cr
TF covariance                           \tab Cv_TaxTb           \tab Cv_T2xT3           \cr
OF covariance                           \tab Cv_OaxOb           \tab Cv_O32xO33         \cr
TMF covariance                          \tab Cv_TMaxTMb         \tab Cv_TM212xTM213     \cr
OMF covariance                          \tab Cv_OMaxOMb         \tab Cv_OM132xOM133     \cr
Residual variance epsilon               \tab eps_imts           \tab eps_2132           \cr
Intercept alpha                         \tab alph_ims           \tab alph_212           \cr
Loading lambda                          \tab lam_ims            \tab lam_212            \cr
Loading delta                           \tab del_ims            \tab del_212            \cr
Loading gamma                           \tab gam_ims            \tab gam_212            \cr
Difference TF                           \tab Dif_Ts             \tab Dif_T2             \cr
Regression intercept (TF)               \tab b0_Ts              \tab b0_T2              \cr
Regression slope beta (TF)              \tab b1_Ts              \tab b1_T2              \cr
Residual variance omega (TF)            \tab omg_Ts             \tab omg_T2             \cr
Difference TMF                          \tab Dif_TMims          \tab Dif_TM212          \cr
Regression slope beta (TMF)             \tab b1_TMims           \tab b1_T212            \cr
Residual variance omega (TMF)           \tab omg_TMims          \tab omg_T212           \cr
Commonality TF                          \tab Com_Ts             \tab Com_T2             \cr
Fixed situation specificity TF          \tab SitSp_Ts           \tab SitSp_T2           \cr
Commonality TMF                         \tab Com_TMims          \tab Com_TM212          \cr
Fixed situation specificity TMF         \tab SitSp_TMims        \tab SitSp_TM212        \cr
Observed variable reliability           \tab RelY_imts          \tab Rel(variable name) \cr
Observed variable consistency           \tab ConY_imts          \tab Con(variable name) \cr
Observed variable specificity           \tab SpeY_imts          \tab Spe(variable name) \cr
}
Note that for difference and commonality variables the fixed reference
situation is always s = 1.
}
\examples{
modelfit <- mmLSTrf(mmLSTrf_exampledata, nSit=2, nTime=3, nMth=2, structural="TF", 
                    includeOMF=TRUE, lat.cov=list(TFcov=FALSE, OFcov=FALSE, 
                    TMFcov=FALSE, OMFcov=FALSE), meanstructure=TRUE, 
                    meas.invar="scalar.b", equiv.ass=list(TF="ess.par", 
                    OF="ess.equiv", OMF="ess.equiv"), addsyntax="")
print(modelfit)
}
\references{
Hintz F., Geiser C., & Shiffman S. (2019). A latent state-trait 
model for analyzing states, traits, situations, method effects, and their 
interactions. Journal of Personality (1-21). https://doi.org/10.1111/jopy.12400
}
\author{
Tinhof Dora
}
