% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{mmLSTrf_exampledata}
\alias{mmLSTrf_exampledata}
\title{Dataset mmLSTrf_exampledata.}
\format{
A data frame with 500 rows and 36 variables
}
\description{
A simulated dataset. The variables are:
}
\details{
\itemize{
  \item Y_1111.
  \item ...
  \item Y_3232
}
}
\keyword{datasets}
