% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{mmLSTrf_RealDataExample}
\alias{mmLSTrf_RealDataExample}
\title{Dataset mmLSTrf_RealDataExample}
\format{
A data frame with 425 rows and 24 variables
}
\description{
This dataset is a subsample from a larger longitudinal study assessing the 
Big Five personality traits in both offline (real-world) and online (digital) 
contexts. Data were collected using the self- and other-rating versions of 
the Big Five Inventory-2 (BFI-2; Soto and John, 2017b) at two measurement 
occasions, with an average interval of 11 days.
}
\details{
From this broader study, six self-rated items measuring the dimension of 
Negative Emotionality (N) were selected. These chosen items correspond to 
those of the short version of the BFI-2 (BFI-2-S; Soto and John, 2017a) and 
are rated on a five-point Likert scale. Three of the items are true-keyed, 
and three are false-keyed.

The six items are listed below; the false-keyed are marked with “*”.

\itemize{
  \item NAt. Worries a lot. 
  \item NDt. Tends to feel depressed, blue. 
  \item NVf. Is emotionally stable, not easily upset*. 
  \item NAf. Is relaxed, handles stress well*. 
  \item NDf. Feels secure, comfortable with self*. 
  \item NVt. Is temperamental, gets emotional easily. 
}

Each item exists in four versions in the dataset, resulting in a total 
of 24 variables. Items assessing Negative Emotionality in the offline 
context are labeled with the prefixes “Of1_” (measurement occasion 
one) and “Of2_” (measurement occasion two), while those in the online 
context are labeled “On1_” and “On2_” respectively.

The resulting dataset reflects a study design with two fixed situations 
(offline and online), two measurement occasions, two measurement methods 
(true- and false-keyed), and three indicators per method (true-keyed: 
NAt, NDt, NVt; false-keyed: NAf, NDf, NVf).

Note: The dataset contains missing values.
}
\keyword{datasets}
