% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{multistate}
\alias{multistate}
\title{Dataset multistate.}
\format{
A data frame with 400 rows and 4 variables
}
\description{
A simulated dataset. The variables are:
}
\details{
\itemize{
  \item y11.
  \item y21.
  \item y12.
  \item y22.
}
}
\keyword{datasets}
