% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{multistate02}
\alias{multistate02}
\title{Dataset multistate02.}
\format{
A data frame with 1000 rows and 9 variables
}
\description{
A simulated dataset. The variables are:
}
\details{
\itemize{
  \item y11.
  \item y21.
  \item y31.
  \item y12.
  \item y22.
  \item y32.
  \item y13.
  \item y23.
  \item y33.
}
}
\keyword{datasets}
