% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{d_taucong}
\alias{d_taucong}
\title{Dataset d_taucong.}
\format{
A data frame with 300 rows and 3 variables
}
\description{
A simulated dataset. The variables are:
}
\details{
\itemize{
  \item y1.
  \item y2.
  \item y3.
}
}
\keyword{datasets}
