% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsttheory_es.R
\name{lsttheory_es}
\alias{lsttheory_es}
\title{Estimate latent state-trait models for experience sampling data}
\usage{
lsttheory_es(
  model,
  ntimepoints,
  nperiods = 1,
  data,
  addsyntax = "",
  la_t_equiv = NULL,
  la_o_equiv = NULL,
  la_s_equiv = NULL,
  vzeta_eqiv = NULL,
  veps_equiv = NULL,
  vtheta_equiv = NULL,
  nu_equiv = NULL,
  alpha_equiv = NULL,
  mtheta_equiv = NULL,
  gamma_t_equiv = NULL,
  manifest_thetacovariates = NULL,
  ...
)
}
\arguments{
\item{model}{integer or character. Can be 1-9 or one of the following: c("MSSTinvar", "STARinvar", "PTARinvar", "ITARinvar", "PITARinvar", "STAR2", "PTAR2", "ITAR2", "PITAR2"). Each model implies invariance assumptions, but details can be changed with the options below.
\enumerate{
\item Model 1 ("STinvar" or "MSSTinvar") is a multistate-singletrait model (no autoregression). Models 1-5 assume state (residual)- and trait-equivalence, i.e., all factor loadings are fixed to 1 and all intercepts are fixed to 0.
\item Model 2 ("STARinvar") is a singletrait model with autoregression.
\item Model 3 ("PTARinvar") is a model with period-specific (usually day-specific) traits. 
\item Model 4 ("ITARinvar") is a model with indicator-specific traits.
\item Model 5 ("PITARinvar") is a model with period- and indicator-specific traits.
\item Models 6 - 9 ("STAR2", "PTAR2", "ITAR2", "PITAR2") are the same as models 2-5, but with state (residual)-congenericity and measurement invariance within periods. In experience sampling data, periods usually correspond to days. 
}}

\item{ntimepoints}{integer. The total number of measurement occasions on which data was collected.}

\item{nperiods}{integer. The number of periods (e.g. days or weeks) on which data was collected.}

\item{data}{a data.frame. This data frame contains the observed variables, sorted by time t and then 
by indicator i, i.e., Y11, Y21, Y31, ... Y12, Y22, Y32 ... Y15, Y25, Y35 ... etc.}

\item{addsyntax}{character string. Will be added to generated lavaan syntax.}

\item{la_t_equiv}{Character. Invariance option for factor loadings of the latent trait. 
Can be one of \code{c("one", "period.invar", "free")}.}

\item{la_o_equiv}{Character. Invariance option for factor loadings of the occasion factor (OCC). 
Can be one of \code{c("one", "invar", "period.invar", "free")}.
In models without autoregression this corresponds to the factor loadings 
of the state variables (for models with a single trait or period-specific traits) or state residual variables (for models with indicator-specific traits).}

\item{la_s_equiv}{Character. Invariance option for autoregression between occasion factors. Can be one of \code{c("zero", "invar", "interval.invar", "free")}.}

\item{vzeta_eqiv}{Character. Invariance option for variances of the state residual (zeta) variables. Can be one of \code{c("invar", "period.invar", "free")}.}

\item{veps_equiv}{Character. Invariance option for variances of the residual (epsilon) variables. 
Can be one of \code{c("invar",  "time.invar", "indicator.invar", "period.invar", "free")}.}

\item{vtheta_equiv}{Character. Invariance option for variances of the latent trait. Can be one of \code{c("invar","indicator.invar", "free")}.}

\item{nu_equiv}{Character. Invariance option for intercepts of the indicators. Can be one of \code{c("zero","period.invar", "free")}.}

\item{alpha_equiv}{Character. Invariance option for intercepts of the latent states. 
Only relevant for models with a single trait or period-specific traits.
Can be one of \code{c("zero","period.invar", "free")}.}

\item{mtheta_equiv}{Character. Invariance option for means of the latent traits. Character. Can be one of \code{c("invar","indicator.invar", "free")}.}

\item{gamma_t_equiv}{Character. Invariance option for regression coefficient from covariates to the latent trait variables.}

\item{manifest_thetacovariates}{Vector or single character. Name or the variable (or variables) in the dataset which are covariates that further explain the trait variables in the model.}

\item{...}{Further arguments passed to lower-level functions}
}
\description{
Estimate latent state-trait models for experience sampling data
}
\author{
Julia Norget
}
